+++ blocklets/ai-studio/api/src/store/migrations/20240529193034-add-appearance.ts
@@ -1,0 +1,11 @@
+import { DataTypes } from 'sequelize';
+
+import type { Migration } from '../migrate';
+
+export const up: Migration = async ({ context: queryInterface }) => {
+  await queryInterface.addColumn('Projects', 'appearance', { type: DataTypes.JSON });
+};
+
+export const down: Migration = async ({ context: queryInterface }) => {
+  await queryInterface.removeColumn('Projects', 'appearance');
+};