+++ blocklets/ai-studio/src/pages/project/settings/appearance-setting.tsx
@@ -1,22 +1,11 @@
+import { UpdateProjectInput } from '@api/routes/project';
 import { useLocaleContext } from '@arcblock/ux/lib/Locale/context';
 import { SaveRounded } from '@mui/icons-material';
 import { LoadingButton } from '@mui/lab';
-import { Box, Dialog, DialogContent, Divider, Stack, Typography, dialogContentClasses, styled } from '@mui/material';
-import { bindDialog, usePopupState } from 'material-ui-popup-state/hooks';
-import { useEffect, useState } from 'react';
-import { ChromePicker } from 'react-color';
+import { Box, Stack, Typography } from '@mui/material';
 
-const defaultColors: string[] = [
-  '#ffffff',
-  '#F47373',
-  '#697689',
-  '#37D67A',
-  '#2CCCE4',
-  '#555555',
-  '#dce775',
-  '#ff8a65',
-  '#ba68c8',
-];
+import FontFamilySetting from './font-family-setting';
+import PrimayColor from './primary-color';
 
 export default function AppearanceSetting({
   value,
@@ -25,101 +14,39 @@ export default function AppearanceSetting({
   readOnly,
   submitLoading,
 }: {
-  value: string | undefined;
+  value: UpdateProjectInput;
   onSubmit: () => void;
   set: (key: string, value: any) => void;
   readOnly: boolean;
   submitLoading: boolean;
 }) {
-  const [selectedColor, setSelectedColor] = useState<string | undefined>(value || defaultColors[0]);
   const { t } = useLocaleContext();
-  const dialogState = usePopupState({ variant: 'dialog' });
-
-  const handleChangeComplete = (color: { hex: string }) => {
-    setSelectedColor(color.hex);
-  };
-
-  const handleClick = (color: string) => {
-    setSelectedColor(color);
-  };
-
-  useEffect(() => {
-    set('primaryColor', selectedColor);
-  }, [selectedColor]);
 
   return (
-    <Box>
-      <Typography variant="subtitle2" mb={0.5}>
-        {t('primaryColor')}
-      </Typography>
-      <Stack gap={2}>
-        <Stack direction="row" gap={1} alignItems="center">
-          {defaultColors?.map((color) => (
-            <Box key={color} border={color === selectedColor ? '1px solid #030712' : ''} borderRadius="4px">
-              <ColorBox bgcolor={color} onClick={() => handleClick(color)} />
-            </Box>
-          ))}
-
-          <Divider orientation="vertical" flexItem sx={{ mx: '2px' }} />
-          <Box onClick={dialogState.open} sx={{ position: 'relative' }}>
-            <ChromePickerBox bgcolor={selectedColor} />
-          </Box>
-        </Stack>
-        <Box>
-          <LoadingButton
-            disabled={readOnly}
-            loading={submitLoading}
-            variant="contained"
-            loadingPosition="start"
-            startIcon={<SaveRounded />}
-            onClick={onSubmit}>
-            {t('save')}
-          </LoadingButton>
-        </Box>
-      </Stack>
-
-      <Dialog
-        {...bindDialog(dialogState)}
-        hideBackdrop
-        sx={{
-          mt: '90px',
-          [`.${dialogContentClasses.root}`]: {
-            padding: '8px !important',
-          },
-        }}
-        PaperProps={{
-          elevation: 0,
-          style: {
-            position: 'absolute',
-            top: '152px',
-            right: '130px',
-            margin: 0,
-          },
-        }}>
-        <DialogContent>
-          <Box mt={2}>
-            <ChromePicker onChangeComplete={handleChangeComplete} color={selectedColor} disableAlpha />
-          </Box>
-        </DialogContent>
-      </Dialog>
-    </Box>
+    <Stack gap={2}>
+      <Box>
+        <Typography variant="subtitle2" mb={0.5}>
+          {t('primaryColor')}
+        </Typography>
+        <PrimayColor value={value} set={set} />
+      </Box>
+      <Box>
+        <Typography variant="subtitle2" mb={0.5}>
+          {t('fontFamily')}
+        </Typography>
+        <FontFamilySetting set={set} value={value} />
+      </Box>
+      <Box>
+        <LoadingButton
+          disabled={readOnly}
+          loading={submitLoading}
+          variant="contained"
+          loadingPosition="start"
+          startIcon={<SaveRounded />}
+          onClick={onSubmit}>
+          {t('save')}
+        </LoadingButton>
+      </Box>
+    </Stack>
   );
 }
-
-const ColorBox = styled(Box)({
-  boxSizing: 'border-box',
-  width: 20,
-  height: 20,
-  borderRadius: 4,
-  margin: 2,
-  border: '1px solid #d6d6d6',
-});
-
-const ChromePickerBox = styled(Box)({
-  boxSizing: 'border-box',
-  width: 24,
-  height: 24,
-  borderRadius: 4,
-  margin: 2,
-  border: '1px solid #d6d6d6',
-});